chore(fc-pallet-orders): types -> types.rs
